requirements.txt
behave==1.2.5
enum34==1.0.4
parse==1.6.6
parse-type==0.3.4
PyHamcrest==1.8.5
selenium==2.48.0
six==1.10.0
wheel==0.24.0
